make more sensible structure
